fixed setting configuration.
