+++ Eyes.Images.DotNet/Images/Eyes.cs
@@ -71,7 +71,7 @@ namespace Applitools.Images
             return new Configuration(configuration_);
         }
 
-        public void SetConfiguration(IConfiguration configuration)
+        protected override void SetConfigImpl(IConfiguration configuration)
         {
             configuration_ = new Configuration(configuration);
         }
