+++ Eyes.LeanFT.DotNet/Eyes.cs
@@ -66,7 +66,7 @@
             return new Configuration(configuration_);
         }
 
-        public void SetConfiguration(IConfiguration configuration)
+        protected override void SetConfigImpl(IConfiguration configuration)
         {
             configuration_ = new Configuration(configuration);
         }
