+++ Eyes.Windows.DotNet/EyesWindowsBase.cs
@@ -71,7 +71,7 @@ namespace Applitools
             return new Configuration(configuration_);
         }
 
-        public void SetConfiguration(IConfiguration configuration)
+        protected override void SetConfigImpl(IConfiguration configuration)
         {
             configuration_ = new Configuration(configuration);
         }
